+++ lib/src/pagination/fixed_size_page.dart
@@ -1,7 +1,7 @@
-import 'package:json_api/src/pagination/pagination.dart';
+import 'package:json_api/src/pagination/pagination_strategy.dart';
 import 'package:json_api/src/query/page.dart';
 
-class FixedSizePage implements Pagination {
+class FixedSizePage implements PaginationStrategy {
   final int size;
 
   FixedSizePage(this.size) {