+++ lib/src/query/include.dart
@@ -2,7 +2,17 @@ import 'dart:collection';
 
 import 'package:json_api/src/query/query_parameters.dart';
 
+/// Query parameter defining inclusion of related resources.
+/// @see https://jsonapi.org/format/#fetching-includes
 class Include extends QueryParameters with IterableMixin<String> {
+  /// Example:
+  /// ```dart
+  /// Include(['comments', 'comments.author']).addTo(url);
+  /// ```
+  /// encodes into
+  /// ```
+  /// ?include=comments,comments.author
+  /// ```
   Include(Iterable<String> resources)
       : _resources = [...resources],
         super({'include': resources.join(',')});