+++ lib/src/query/page.dart
@@ -1,9 +1,17 @@
 import 'package:json_api/src/query/query_parameters.dart';
 
-/// The "page" query parameters
+/// Query parameters defining the pagination data.
+/// @see https://jsonapi.org/format/#fetching-pagination
 class Page extends QueryParameters {
-  static final _regex = RegExp(r'^page\[(.+)\]$');
-
+  /// Example:
+  /// ```dart
+  /// Page({'limit': '10', 'offset': '20'}).addTo(url);
+  /// ```
+  /// encodes into
+  /// ```
+  /// ?page[limit]=10&page[offset]=20
+  /// ```
+  ///
   Page(Map<String, String> parameters)
       : _parameters = {...parameters},
         super(parameters.map((k, v) => MapEntry('page[${k}]', v)));
@@ -13,6 +21,8 @@ class Page extends QueryParameters {
         ..removeWhere((k, v) => k == null));
 
   String operator [](String key) => _parameters[key];
+
+  static final _regex = RegExp(r'^page\[(.+)\]$');
 
   final Map<String, String> _parameters;
 }