+++ lib/src/query/query_parameters.dart
@@ -8,12 +8,14 @@ class QueryParameters {
 
   bool get isNotEmpty => _parameters.isNotEmpty;
 
+  /// Adds (or replaces) this parameters to the [uri].
   Uri addToUri(Uri uri) => isEmpty
       ? uri
       : uri.replace(queryParameters: {...uri.queryParameters, ..._parameters});
 
-  QueryParameters merge(QueryParameters moreParameters) =>
-      QueryParameters({..._parameters, ...moreParameters._parameters});
+  /// Merges this parameters with [other] parameters. Returns a new instance.
+  QueryParameters merge(QueryParameters other) =>
+      QueryParameters({..._parameters, ...other._parameters});
 
   /// A shortcut for [merge]
   QueryParameters operator &(QueryParameters moreParameters) =>